requirements.txt
Appium-Python-Client
PyYAML
allure-pytest
